Extract date formatting into PostRsDateFormatter
